move confirmation into submodule
